Major change on Links and Markers.
Absolute Links are now created to handle -q binaries. For x86,
calculateDisplacement() now returns unsigned long for Linux kernel.

Bingen is broken due to this change.

+++ src/instr/linked-x86_64.h
@@ -43,7 +43,7 @@ public:
     virtual void accept(InstructionVisitor *visitor) { visitor->visit(this); }
 protected:
     Instruction *getInstruction() const { return instruction; }
-    unsigned calculateDisplacement();
+    unsigned long calculateDisplacement();
     void makeDisplacementInfo();
 };
